alphabetic sorting if not explicit + adding sorting module
