+++ sphinx_gallery/gen_gallery.py
@@ -19,6 +19,7 @@ from . import glr_path_static
 from .gen_rst import generate_dir_rst, SPHX_GLR_SIG
 from .docs_resolv import embed_code_links
 from .downloads import generate_zipfiles
+from .sorting import ExplicitOrderKey
 
 try:
     FileNotFoundError
@@ -313,22 +314,3 @@ def setup(app):
 
     app.connect('build-finished', sumarize_failing_examples)
     app.connect('build-finished', embed_code_links)
-
-
-class ExplicitOrderKey(object):
-    """Explicitly order a list of strings."""
-    def __init__(self, ordered_list):
-        if not isinstance(ordered_list, list):
-            raise ValueError('ordered_list must be a list')
-        self.ordered_list = ordered_list
-        self.unordered_items = []
-
-    def __call__(self, item):
-        if item in self.ordered_list:
-            # Return that item's explicit position
-            return self.ordered_list.index(item)
-        else:
-            # Keep a running list of non-explicitly ordered things
-            # Add any non-specified strings to the end of the sort
-            self.unordered_items.append(item)
-            return len(self.unordered_items) + len(self.ordered_list)
